2-9.注文の際にメールを送る(上級)(注文をする)の非同期処理
