+++ src/main/java/com/example/controller/OrderController.java
@@ -52,40 +52,10 @@ public class OrderController {
         webApiResponseObject.setStatus("success");
         webApiResponseObject.setMessage("OK.");
         webApiResponseObject.setErrorCode("E-00");
-        System.out.println("★終わったよ");
         // メール送信を非同期で行う
-        sendOrderConfirmationEmail(destination, address, order);
-        System.out.println("★メール終わったよ");
+        emailService.sendOrderConfirmationEmail(destination, address, order);
         return webApiResponseObject;
     }
 
-    @Async("taskExecutor")
-    public void sendOrderConfirmationEmail(Destination destination, Address address, Order order) {
-        String subject = "注文完了メール";
-        StringBuilder sb = new StringBuilder();
 
-        // メールに記載するため、注文情報を取得
-        Order orderInfo = emailService.getOrderInfo(order.getId());
-
-        sb.append(destination.getDestinationName()).append("様").append("\n\n");
-        sb.append("ご注文ありがとうございます。\n");
-        sb.append("ご注文内容は以下の通りです。\n\n");
-        sb.append("===============================\n");
-        sb.append("注文商品：\n");
-        List<OrderItem> orderItemList = orderInfo.getItemList();
-        for (OrderItem item : orderItemList){
-            sb.append("　　商品名：").append(item.getItem().getName()).append("\n");
-            sb.append("　　サイズ：").append(item.getSize()).append("\n");
-            sb.append("　　個数：").append(item.getQuantity()).append("点").append("\n");
-            sb.append("　　1点当たりの価格：").append(item.getItem().getPrice()).append("円").append("\n");
-        }
-        sb.append("\n");
-        sb.append("お届け先：").append(address.getPrefecture() + " "+ address.getMunicipalities() + " "+ address.getAddress()).append("\n");
-        sb.append("お届け予定日時：").append(order.getDeliveryDate()).append("\n");
-        sb.append("===============================\n\n");
-        sb.append("またのご利用をお待ちしております。");
-
-        String text = sb.toString();
-        emailService.sendSimpleMessage(destination.getDestinationEmail(), subject, text);
-    }
 }
